Replace module 'countries' by subpackage 'holidays'
