+++ python/holidays/__init__.py
@@ -1,0 +1,8 @@
+from .holidays import Locale
+
+__all__ = [
+    "Locale",
+    "en-US",
+    "hr-HR",
+    "sk-SK",
+]
