+++ python/holidays/hr-HR.py
@@ -1,0 +1,22 @@
+from .holidays import Locale
+
+
+class hr_HR(Locale):
+    u"""
+    01-01: [NF] Nova Godina
+    01-06: [NRF] Bogojavljanje / Sveta tri kralja
+    05-01: [NF] Praznik rada
+    06-22: [NF] Dan antifašističke borbe
+    06-25: [NF] Dan državnosti
+    08-05: [NF] Dan domovinske zahvalnosti
+    08-15: [NRF] Velika Gospa
+    10-08: [NF] Dan neovisnosti
+    11-01: [NRF] Dan svih svetih
+    12-25: [NRF] Božić
+    12-26: [NRF] Sveti Stjepan
+    0 days after Easter: [NRV] Uskrs
+    1 day after Easter: [NRV] Uskršnji ponedjeljak
+    60 days after Easter: [NRV] Tijelovo
+    """
+
+    locale = "hr-HR"
